Extract binary stuff into separate file
